+++ docs/news.rst
@@ -25,7 +25,7 @@ Documentation
 Changed
 ~~~~~~~
 
-- **BACKWARDS-INCOMPATIBLE CHANGE:** The ``IEggStorage.get()`` interface returns a ``(version, filename)`` tuple, instead of a ``(version, file)`` tuple.
+- The ``IEggStorage.get()`` interface can return a ``(version, filename)`` tuple in addition to a ``(version, file)`` tuple.
 - Drop support for end-of-life Python version 3.7.
 
 Web UI